Fetch using http
Rust is unable to access HTTPS without 3rd party libraries written
in other languages. So, that requires either a C compiler to be
present, or common libraries such as OpenSSL to be installed.

And it both cases, people complain. So, until their language
supports HTTPS, fall back to using HTTP to fetch remote files.

Fortunately, all the files we download include signatures that are
verified.